Integrated configurator with a VERY basic UI! Restored original joystick precision, and fixed joystick calibration to enable full values in any direction without pressing harder than what is expected against the gate.

+++ main/hall_joystick.cpp
@@ -26,7 +26,8 @@ void HallJoystick::Init() {
 }
 
 int HallJoystick::Normalize(int val, const Bounds& in) {
-  int mapped = (val - in.min) / (in.max - in.min) * (out_.max - out_.min) + out_.min;
+  int mapped = static_cast<float>(val - in.min) /
+    static_cast<float>(in.max - in.min) * (out_.max - out_.min) + out_.min;
   return constrain(mapped, out_.min, out_.max);
 }
